msgd: handle abrupt restart of remote node [#2840]
Sometimes when a remote node restarts abruptly, queues which were created on
that node, are unable to be opened again when that node comes up.

There is a race condition when the remote node goes down between msgd getting
the CLM and MDS events indicating node down, and immd removing the implementer
for that remote node. When msgd gets the CLM and MDS events indicating node
down it temporarily becomes the implementer for any queues on that node so that
it can remove the entries in IMM. If IMM has not yet removed the implementer,
msgd will fail to remove the IMM entries. When the remote node comes back up,
and the queues are opened, they will fail because the IMM entries are still
there.

When msgd recevies ERR_EXIST from implementer set in this case, it should
treat it as TRY_AGAIN.