clmd: Increase message priority of CLMSV_CLMS_MDS_NODE_EVT to be sent to main thread [#2842]
In the event of stop/start standby controller, the node is stopped that
generates the MDS event CLMSV_CLMS_MDS_NODE_EVT. This event is being sent
to main thread with NORMAL priority. When the node is started again, the
other event like CLMSV_CLUSTER_JOIN_REQ is being sent with HIGH priority.

The race happens as CLMSV_CLMS_MDS_NODE_EVT is processed after the event
CLMSV_CLUSTER_JOIN_REQ, possibly caused by the priority.

The patch sets priority of CLMSV_CLMS_MDS_NODE_EVT as high as the others
so that the order of messages processed in main thread should depend on
the timing order of events that occurred.