smfd: Remove duplicate DU/AU on SU/comp level in one-step upgrade [#2227]
This fix extends, and also improves the previous one of #2209 which was
done for node level DU/AU. In particular, it eliminates any duplicate AU/DU
on SU and component level when merging forAddRemove and forModify/Rolling
procedures into a single-step procedure for one-step upgrade execution mode.

Improvement is made to remove units that have duplicates within the
forAddRemoveDU/AU lists themselves, but not having duplicate in the merged
symmetric DU/AU lists, which is the case that #2209 failed to consider.