ntf: Don't use old style SaNameT with 255 chars DN in ntfimcnd [#2850]
osafntfimcnd uses old style SaNameT when sending notification. The
SaNameT's length will be adjusted to 256 when its value contains
255 characters excluded the NULL character. This SaNameT is now
not a valid extended name and causes osafntfimcnd to exit.

This patch helps avoid the above issue by using the new style of
SaNameT if the SaNameT's value is a 255 chars DN.