msgd: put node down handling on thread [#2852]
If multiple nodes go down simultaneously which are hosting msg queues (e.g.
multiple VMs on a host, and the host goes down), msgd can take a long time to
process the node downs which blocks the main thread, and therefore the
healthcheck doesn't get processed, so msgd dies, which restarts the controller.

msgd needs to sit in a loop waiting for imm to release the implementers for
each of the down nodes. For many nodes which went down simultaneously this can
take up to 20 seconds when done serially.

Node down logic needs to be put on a thread, so that we can continue to process
other messages like healthcheck. This also allows us to parallelize the node
down handling.