ntf: fix ntfimcnd issue with empty SaNameT [#2854]
ntfimcnd unexpectedly exits when reserving memory for empty SaNameT
attribute. The issue happens becasue ntfimcnd passes the SaNameT's
length which is 0, to saNtfPtrValAllocate function.

This patch helps avoid the issue by increasing the SaNameT' length
to 1 before passing it to the saNtfPtrValAllocate function. A test
case to verify the issue is also included in this patch.