base: Place TraceLog instance on heap memory V3 [#2860]
When a process calls exit(), the exit_handler trigger
__do_global_dtor_aux then calls TraceLog destructor.
One of process thread is calling TraceLog::Log while
the destructor is called. This leads to a coredump.
The process (which could by any applications) calling
exit() first is responsible to make exit() thread-safe.
However, the TraceLog should also avoid the coredump
occurring at OpenSAF side.

Runtime allocate TraceLog pointer so that its memory
is located in heap. This allocation to avoid its
desctructor to be called as part of __do_global_dtor_aux.