ntf: restart ntfimcnd if operation invoker name is missing [#2859]
If ntfimcnd was restarted during a CCB, it might receive ObjectDelete
operation as the first event. ccbUtilCcbData, which was initialized
in this case, does not contain the operation invoker name. This causes
ntfimcnd to crash and generates a coredump when the ccbUtilCcbData
being used in the next ApplyCallback.

This patch helps avoid the above issue by validating the operation
invoker name in ApplyCallback and restarting ntfimcnd if the invoker
name is invalid. An error report notification will be sent out when
ntfimcnd being successfully started.