log: fix restoring ref counter for deleted stream [#2870]
In the methods LogAgent::saLogStreamClose() and LogAgent::saLogWriteLogAsync(),
the client is deleted if failed to recover; however, the pointer to the log
stream of this client has not been reset. Therefore, when the destrustor of
ScopeData runs, the reference counter could be restored on deleted log stream.

This patch resets the pointer to nullptr in such cases. With that, the ScopeData
won't run `RestoreRefCounter()` on the deleted stream.