imm: prioritize to elect IMMND on the active node [#2862]
The coordinator IMMND on PL-3 was crashed, the active IMMD then elected the new
coordinator on the standby node, SC-2, but failed because IMMND on the SC-2 was
restarted also. As the result, the active IMMD exited and failure-over
happened. After that, SC-2 took active role and found no candidate for new IMMND
coordinator, so cluster was rebooted.

We can prevent this happen if the active IMMD prioritizes to elect the
coordinator which is located on the same site with himself if the IMMND database
is up-to-date.